requirements.txt
google-adk[extensions]
playwright
pydantic
aiofiles